Profile Pictures: Refactor upload handling code
Precursor to #71
Should prevent problems in #15

This patch refactors a fair bit of code surrounding the way user profile
pictures are uploaded and saved. The existing process is full of nested
blocks and hard-coded information, preventing some flexibility in the
way we deal with uploads.

The refactor should provide the following benefits:

 - Image types are recognized from their actual data and not from the
   file extension. This closes off the possibility to defeat the rule by
   simply changing the filename. The new check relies on the built-in
   `imghdr` library in Python

 - Some image operations have been modularized in order to isolate and
   simplify their behaviors. This should allow incorporating
   alternatives to the S3 storage, such as for when developing locally.

These changes alter the way the image filenames are generated.
Previously, the stored filename was a function of the first fifty
characters in the uploaded filename. Now, the filename is generated as a
function of the SHA1 hash of the uploaded file contents, which not only
greatly limits the range of characters sent as the filename (no possible
encoding issues, for example), but also ends up mapping identical images
to identical filenames, regardless of the given name.

This last point about mapping to the same name could carry a negative
implication: we could end up having two references to the same file and
accidentally delete "both" because we deleted the singular mapping
endpoint. It does not appear like this will be a problem given the way
the app is structured.

_Example_

 - User uploads profile picture
 - User uploads picture for some other use, but the contents are
   identical to the profile picture
 - User deletes this picture, but since it shares the same filename as
   the profile picture (they are identical), both get erased.

This new file naming behavior is more regular and standard than the
existing formula, but we could just as easily reproduce the old behavior
by updating the `image_to_user_filename()` function.