Delay startup of web app to allow DB to initialize
Previously we have had some sporadic issues where the Flask app was not
properly starting up because it could not attach to the database.

I believe that this problem was due to timing issues initializing
postgres.

Now, the Flask app startup is delayed by five seconds to allow proper DB
initialization before starting and this seems to be resolving any of
those sporadic issues.