Automate building minified bundle
This commit brings in Flask-Webpack which is used to link to specific
versions of the bundled JavaScript. Because it generates files with a
hash in them, these bundles can be cached indefinitely, while
Flask-Webpack makes sure that we load the appropriate ones.

I have modified the `Procfile` and `Dockerfile` to force a rebuild of
the JavaScript when starting the server to make sure that the manifest
file and bundles are ready.