Rearrange and modularize debugging code
In the first commit for this branch, I added some direct connection code
to hook into the PyCharm debugger. My goal is to enable connecting to
one of any given number of debuggers before merging into `master`

It is a bit unfair to deploy a release where a feature is only available
in a licensed commercial software package (PyCharm).

The next step will thus be to try and connect into a Sublime Text
debugger. The changes in this patch setup a system for connecting to the
first of many potential debugging endpoints.