Add production/development status to config and safeguard deployment
Previously the app performed no checking and recorded no state of
whether or not we were running in production or locally while
developing/testing. Unfortunately, I accidentally ran the local
development server with the production config, which led to the
initialization script connecting to the production database, wiping it,
and loading random data into it.

This patch adds a new variable into the config to explicitly specify
which environment is running and adds some safeguards to prevent
accidentally doing what I did. It will not run the production script
without `config.APP_ENV == 'production'` nor will it run the local
development script _with_ `config.APP_ENV == 'production'`.

While it will still be possible to wipe the database or connect
improperly, this should protect against the accidental case I ran into.