Fix: Production had a JS error not present in development (#149)
There was an error when trying to delete previously saved searches. This
error was only present in production. The "cause" was the `-p` setting
to webpack, which minified the bundle. By removing this and shipping the
unminified code we have resolved the bug and moved from a bundle of 60
KB to one of 120 KB

This shall be addressed solidly in another PR.