Adds Angular dependency-injection annotations to allow minification (#151)
Resolves #150

Angular has a runtime dependency-injection system based on the names of
function parameters. Unfortunately, in the process of minification where
those parameters are rewritten in order to shrink the app size, Angular
loses the link it formerly had to know which variables to send as input.

In this patch we add annotation where necessary with the help of the
`ng-annotate` package. It can both add the annotations directly into a
file in the source code as well as provide for some compile-time
assistance through the `'ngInject';` directive inside things like
anonymous functions. The compile-time process is handled through the
webpack loader.

This patch should eliminate the bugs raised in production that were
temporarily fixed in #149 by removing the minification from the process.