Optimize Dockerfile and use Dockerhub image (#163)
Previously, the Dockerfile wasn't well optimized. It required about 400 MB of downloads and took between eight minutes and half an hour to build.

Now, the Dockerfile itself has been updated to minimize the number of introduced layers in the unionfs of the build, plus it has been uploaded to Docker Hub so that we can download the pre-built image and work with that instead of needing to rebuild it locally.