Fix bug uploading pictures (#166)
Previously, in #110, I refactored the picture-handling code to clean it
up. In the process I added `if user_id is not user.id` when that was
wrong. This `is` checked identity of integers, which would not be equal
in normal cases.

Now, this is fixed by replaced `is not` with `!=` and should fix the
authentication problem.