Fix/missing manifest (#173)
Previously there was a bug wherein running the local server Docker image could fail if webpack had not previously run, producing the manifest.json file. Flask was failing because it expected to find that file in order to build appropriate JavaScript and CSS URLs.

Now, the local test server will wait until it finds a manifest file to initialize the Flask app.