Create basic page tracking API (#183)
* Create basic page tracking API

We currently have very little knowledge about how people use the
Communityshare website. This includes metrics such as visits and unique
visitors and it also includes broader information such as how people
navigate the site.

This patch creates a new API endpoint which we can hit to start tracking
information about page views. It stores a user id and the paths the user
routes to from the path he or she left. With a timestamp logging this
information, we can start to analyze the data to see which parts of the
page are serving our users.

The analytics page view requests should be "fire-and-forget", meaning
that there should be no required handling of the response in the client.

At best, we could attempt to detect a failure, but reasonably speaking
we can ignore the response.

Correspondingly there is no meaningful information that the server ought
to return from such a request other than at most whether it succeeded.