Log path changes to the /analytics/views/ endpoint (#184)
This PR triggers the app to start taking advantage of the new tracking API brought in in #183. It will record an activity log of users who visit the site and start to provide insight into how people are using it and navigating around. We will use this information to guide development on the site in order to better and more effectively meet our user needs.