REST API: Add `/users/` endpoint (#174)
* REST API: Add `/users/<id>` and `/users/search` endpoints

This patch continues the "inspiration" work on the REST API to explore
how we can reorganize the API into a more RESTy and bandwidth-friendly
nature.

Significant refactoring was also performed on the `/users/` endpoint to
decouple it from the existing user search function in the `Users` model.
Pulling the actual queries into the endpoint modules makes the queried
material more explicit and allow a bit more freedom to trim the queries
and data returned.

I have experimented with HATEOAS concepts here returning links for users
in the search endpoint. This is convenient in standard REST API clients
which can easily follow those links for more information. Although the
links add considerable text content, because they are mostly repeated
information, the byte-size after compression is significantly smaller.