Add remaining-character-count counter to bio on edit settings page (#230)
Resolves #99

This is just a simple counter and it doesn't prevent trying to save the
form when more than 1000 characters are entered. Resolving that issue is
a bigger challenge and will taken care of in another PR

**Testing**

Navigate to the Edit Profile page and write in the bio box. Notice that
it should show how many more characters are eligible to be typed in
while still being able to save to the server.