Create non-admin test users for local development (#235)
Previously we had a test admin user in the local dev environment
`admin@example.com` but no non-admin users.

Now, we create two new users, `teacher@example.com` and
`partner@example.com` with passwords `teacher` and `parter`,
respectively. These are non-admin users.

Please note, these users are not currently setup as teachers and
partners because of additional work required to make that happen. This
is the plan however and will be handled in a future PR.