Replace jsonify with response sub class (#245)
This removes flask.jsonify from everywhere but app.py, where it is applied to every dict response that passes by.

The trickiest part of this is the Response sub-class that overrides force_type to add dict-to-json conversions. See Response.force_type documentation for more details. The pattern here is fairly general: If we find we are writing boilerplate but can't find a good way to eliminate it with the flask hooks, there may be an opportunity to do so by adding a default or overriding a method on Response.

#### Testing
Relying on the test suite here. If this doesn't work it breaks essentially every endpoint, so tests provide good if coarse signal.